% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_impute_opt}
\alias{parse_impute_opt}
\title{parse options in the fastphase model}
\usage{
parse_impute_opt(filename)
}
\arguments{
\item{filename}{path to binary file from impute command}
}
\description{
parse options in the fastphase model
}
