% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_impute_par}
\alias{parse_impute_par}
\title{parse parameters in the fastphase model}
\usage{
parse_impute_par(filename, ic = -1L)
}
\arguments{
\item{filename}{path to binary file from impute command}
}
\description{
parse parameters in the fastphase model
}
