% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_joint}
\alias{parse_joint}
\title{parse parameters in joint model}
\usage{
parse_joint(filename)
}
\arguments{
\item{filename}{path to binary file from joint command}
}
\description{
parse parameters in joint model
}
