% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_joint_post}
\alias{parse_joint_post}
\title{parse posterior probabilty in joint model}
\usage{
parse_joint_post(filename, chunk = 0L)
}
\arguments{
\item{filename}{path to binary file from joint command}

\item{chunk}{which chunk to extract}
}
\description{
parse posterior probabilty in joint model
}
