% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseless-package.R
\docType{package}
\name{phaseless-package}
\alias{phaseless}
\alias{phaseless-package}
\title{Imports}
\description{
Admixture and imputation for low coverage sequencing data in one goal.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Zilong-Li/phaseless}
  \item Report bugs at \url{https://github.com/Zilong-Li/phaseless/issues}
}

}
\author{
\strong{Maintainer}: Zilong Li \email{zilong.dk@gmail.com} (\href{https://orcid.org/0000-0001-5859-2078}{ORCID}) [copyright holder]

}
